requirements.runner.txt
# DB/ETL
pandas==1.5.3
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9

# Binance + HTTP
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.2
aiohttp==3.10.11
requests>=2.31
tqdm==4.67.1
PyYAML==6.0.2
numpy<2

tenacity>=8.2
loguru==0.7.2
pydantic<2

# Log
python-json-logger==2.0.7
